revert ownerships to Debian defaults (#12)
In the Debian package configuration, configuration files are owned by
root. They are also world-readable (which is not ideal), so we keep
the "owned by barman group" permission, but remove the "owned by
barman user" permission.

In other words, barman shouldn't be able to modify its own
configuration files. Just a precaution.

Closes: #11